requirements.txt
Django==4.2
django-environ==0.10.0
django-cors-headers==3.14.0
djangorestframework==3.14.0
Pillow==9.5.0
django-storages==1.13.2
django-ckeditor==6.5.1


djangorestframework-simplejwt==5.2.2
social-auth-app-django==5.2.0
social-auth-core==4.4.2
argon2-cffi==21.3.0
djoser==2.2.0
PyJWT==2.7.0

python-slugify==8.0.1
requests==2.31.0

python-dotenv==1.0.0
coverage==7.2.5
psycopg2-binary==2.9.6

gunicorn==20.1.0
whitenoise==6.4.0
dj-database-url==2.0.0

boto3==1.26.142
django-storages==1.13.2